Update README.md
Explaining how codebases can evolve in real time, this detail wasn't clear enough.

For example- Tell llm "Create a docker image for a debian environment"

Or deploy a codebase and configure your CICD. Building it this way, you will be able to declare software state like Terraform and watch it rollout in real time, with retries i.e crash-proof.

This is a boundless paradigm, and will take time for adoption. Feel free to contribute.